requirements.txt
beautifulsoup4
bs4
fastapi
frontend
nltk
numpy
opencv-python
Pillow
PyMuPDF
PyPDF2
pytesseract
python-multipart
regex
requests
scikit-learn
tensorflow
tensorflow-hub
uvicorn